applied PR comments
